InComNet/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
ftfy==6.2.3
idna==3.10
imageio==2.35.1
joblib==1.4.2
numpy==1.24.4
open-clip-torch-any-py3==1.3.2
opencv-python==4.10.0.84
packaging==24.2
pillow==10.4.0
regex==2024.11.6
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
threadpoolctl==3.5.0
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
